server/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
dnslib==0.9.23
cryptography==41.0.3
requests==2.31.0
PyYAML==6.0.1
dnspython==2.4.2
gunicorn==21.2.0
Werkzeug==2.3.7
